Refactor qp-klp
Refactor qp-klp to accept amplicon-pipeline in the future.
Clean up and capture all Errors so they can correctly be passed back as
success = False and an error-message (No compressed stack-frame msgs).
